feat: Update favorite button color and remove icon position
This commit updates the color of the favorite button in the CharacterInfo screen to #E31B23. It also adjusts the position of the remove icon in the Favorite screen to be positioned at the top right corner of each favorite item. These changes improve the visual appearance and user experience of the application.