feat: Increase image height in Detail screen and update favorite character message
This commit increases the height of the image in the Detail screen from 160 to 190 pixels, improving the visual appearance of the screen. It also updates the message displayed when no favorite characters are found in the Favorite screen to "Favourite character not found !". These changes enhance the user experience of the application.